requirements.txt
streamlit==1.26.0
torch>=1.12.0
torchvision>=0.13.0
numpy>=1.22.0
opencv-python>=4.8.0
Pillow>=9.2.0
